+++ src/device/mtpdevice.cpp
@@ -155,7 +155,7 @@ bool MtpDevice::StartCopy(QList<Song::FileType> *supported_types) {
 static int ProgressCallback(uint64_t const sent, uint64_t const total, void const *const data) {
 
   const MusicStorage::CopyJob *job = reinterpret_cast<const MusicStorage::CopyJob*>(data);
-  job->progress_(static_cast<float>(sent) / total);
+  job->progress_(static_cast<float>(sent) / static_cast<float>(total));
 
   return 0;
